requirements.txt
flask==3.1.0
line-bot-sdk==3.5.1
google-generativeai==0.8.5
google-api-python-client==2.164.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
APScheduler==3.11.0
python-dotenv==1.1.0
gunicorn==23.0.0
